+++ packages/blocks/core/CHANGELOG.md
@@ -15,6 +15,7 @@
 - `Details` block: inner blocks improved.
 - `File` block: inner blocks improved.
 - `Footnote` block: inner blocks improved.
+- `Gallery` block: inner blocks improved.
 
 
 ### Automated Tests